+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Change the verification of the grid in MAPL_GetGlobalHorzIJIndex to avoid collective call
+- Use MAPL_GetGlobalHorzIJIndex to implement MAPL_GetHorzIJIndex
 
 ### Fixed